Add by-the-lake
Instead of embedding the gallery generated by the `makegallery.rb`
command, I'm putting it in the `_data` subdirectory with the same name
as the post. Doing this required some tweaking on the `post_index_entry`
include file so it would pick up either the gallery from the data file
or from the post front matter.

Also needed to clean up the Gemfile and plugins to fully support jekyll
3.x instead of the pre-release.